+++ build.gradle
@@ -1,35 +1,36 @@
 plugins {
-	id 'java'
-	id 'org.springframework.boot' version '3.2.3'
-	id 'io.spring.dependency-management' version '1.1.4'
+    id 'java'
+    id 'org.springframework.boot' version '3.2.3'
+    id 'io.spring.dependency-management' version '1.1.4'
 }
 
 group = 'com.example'
 version = '0.0.1-SNAPSHOT'
 
 java {
-	sourceCompatibility = '20'
+    sourceCompatibility = '20'
 }
 
 repositories {
-	mavenCentral()
+    mavenCentral()
 }
 
 dependencies {
-	// Spring Web & JWT
-		implementation 'org.springframework.boot:spring-boot-starter-web'
-		implementation 'io.jsonwebtoken:jjwt:0.9.1'
+    // Spring Web & JWT
+    implementation 'org.springframework.boot:spring-boot-starter-web'
+    implementation 'io.jsonwebtoken:jjwt:0.9.1'
 
-	// JPA & DB
-		implementation 'org.springframework.boot:spring-boot-starter-data-jpa'
-		compileOnly 'org.projectlombok:lombok:1.18.22'
-		annotationProcessor 'org.projectlombok:lombok:1.18.22'
-		implementation 'javax:javaee-api:8.0'
+    // JPA & DB
+    implementation 'org.springframework.boot:spring-boot-starter-data-jpa'
+    compileOnly 'org.projectlombok:lombok:1.18.22'
+    annotationProcessor 'org.projectlombok:lombok:1.18.22'
+    implementation 'javax:javaee-api:8.0'
+    implementation "com.h2database:h2"
 
-	// Testing
-		testImplementation 'org.springframework.boot:spring-boot-starter-test'
+    // Testing
+    testImplementation 'org.springframework.boot:spring-boot-starter-test'
 }
 
 tasks.named('test') {
-	useJUnitPlatform()
+    useJUnitPlatform()
 }
