+++ src/main/java/sodong/SodongApplication.java
@@ -2,7 +2,10 @@ package sodong;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.boot.autoconfigure.domain.EntityScan;
+import org.springframework.data.jpa.repository.config.EnableJpaRepositories;
 
+@EnableJpaRepositories
 @SpringBootApplication
 public class SodongApplication {
 	public static void main(String[] args) {
