--- src/main/java/sodong/WebSecurityConfig.java
@@ -1,44 +1,0 @@
-//package sodong;
-//
-//import lombok.RequiredArgsConstructor;
-//import org.springframework.boot.autoconfigure.security.servlet.PathRequest;
-//import org.springframework.context.annotation.Bean;
-//import org.springframework.context.annotation.Configuration;
-//import org.springframework.security.config.annotation.web.builders.HttpSecurity;
-//import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
-//import org.springframework.security.config.http.SessionCreationPolicy;
-//import org.springframework.security.crypto.factory.PasswordEncoderFactories;
-//import org.springframework.security.crypto.password.PasswordEncoder;
-//import org.springframework.security.web.SecurityFilterChain;
-//import org.springframework.security.web.authentication.UsernamePasswordAuthenticationFilter;
-//import sodong.security.JwtAuthenticationFilter;
-//import sodong.security.JwtTokenProvider;
-//
-//@RequiredArgsConstructor
-//@EnableWebSecurity
-//public class WebSecurityConfig {
-//
-//    private final JwtTokenProvider jwtTokenProvider;
-//
-//    // 암호화에 필요한 PasswordEncoder 를 Bean 등록합니다.
-//    @Bean
-//    public PasswordEncoder passwordEncoder() {
-//        return PasswordEncoderFactories.createDelegatingPasswordEncoder();
-//    }
-//
-//    @Bean
-//    public SecurityFilterChain securityFilterChain(HttpSecurity http) throws Exception {
-//        return http
-//                .csrf(csrf -> csrf.disable()) // CSRF 보안 토큰 비활성화
-//                .sessionManagement(session -> session.sessionCreationPolicy(SessionCreationPolicy.STATELESS)) // 세션 비활성화
-//                .authorizeRequests((authorizeRequests) ->
-//                        authorizeRequests
-//                                .requestMatchers(PathRequest.toH2Console()).permitAll()
-//                                .requestMatchers("login/**").permitAll()
-//                                .requestMatchers("api/**").hasRole("USER")
-//                                .anyRequest().authenticated()
-//                )
-//                .addFilterBefore(new JwtAuthenticationFilter(jwtTokenProvider), UsernamePasswordAuthenticationFilter.class) // JwtAuthenticationFilter를 UsernamePasswordAuthenticationFilter 전에 넣기
-//                .build();
-//    }
-//}
