+++ src/main/java/sodong/repository/PostingRepository.java
@@ -1,9 +1,7 @@
 package sodong.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
 import sodong.domain.Posting;
 
-@Repository
 public interface PostingRepository extends JpaRepository<Posting, Long> {
 }
